LayoutTests/media/track/track-text-track.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../media-file.js></script>
        <script src=../video-test.js></script>
        <script>            
            
            numberOfTrackTests = 5;
            
            function trackLoaded()
            {
                numberOfTracksLoaded++;
                if (numberOfTracksLoaded == numberOfTrackTests) {
                    testAllTracks();
                }
            }
            
            function testAllTracks()
            {
                findMediaElement();
                var expected = 
                {
                    length : 5,
                    tests:
                    [
                        {
                            property : "kind",
                            values : ["captions", "subtitles", "chapters", "descriptions", "metadata"],
                        },
                        {
                            property : "label",
                            values : ["English Captions", "French Subtitles", "Chapter Tree", "", ""],
                        },
                        {
                            property : "language",
                            values : ["en", "fr", "en", "en", ""],
                        },
                        {
                            property : "cues.length",
                            values : [4, 4, 4, 4, 4],
                        },
                    ],
                };
                testTracks(expected);

                endTest();
            }
            
        </script>
    </head>
    <body>
        <p>Tests TextTrack functionality: kind, label, language, and cues.</p>
        <video>
            <track src="captions-webvtt/tc013-settings.vtt" onload="trackLoaded()" 
                kind="captions" label="English Captions" srclang="en">
            <track src="captions-webvtt/tc013-settings.vtt" onload="trackLoaded()"
                kind="subtitles" label="French Subtitles" srclang="fr">
            <track src="captions-webvtt/tc013-settings.vtt" onload="trackLoaded()"
                kind="chapters" label="Chapter Tree" srclang="en">
            <track src="captions-webvtt/tc013-settings.vtt" onload="trackLoaded()"
                kind="descriptions" label="" srclang="en">
            <track src="captions-webvtt/tc013-settings.vtt" onload="trackLoaded()"
                kind="metadata" label="" srclang="">
        </video>
    </body>
</html>
