LayoutTests/media/track/track-text-track-expected.txt

Tests TextTrack functionality: kind, label, language, and cues.

EXPECTED (tracks.length == '5') OK

*** Testing text track 0
EXPECTED (track.kind == 'captions') OK
EXPECTED (track.label == 'English Captions') OK
EXPECTED (track.language == 'en') OK
EXPECTED (track.cues.length == '4') OK

*** Testing text track 1
EXPECTED (track.kind == 'subtitles') OK
EXPECTED (track.label == 'French Subtitles') OK
EXPECTED (track.language == 'fr') OK
EXPECTED (track.cues.length == '4') OK

*** Testing text track 2
EXPECTED (track.kind == 'chapters') OK
EXPECTED (track.label == 'Chapter Tree') OK
EXPECTED (track.language == 'en') OK
EXPECTED (track.cues.length == '4') OK

*** Testing text track 3
EXPECTED (track.kind == 'descriptions') OK
EXPECTED (track.label == '') OK
EXPECTED (track.language == 'en') OK
EXPECTED (track.cues.length == '4') OK

*** Testing text track 4
EXPECTED (track.kind == 'metadata') OK
EXPECTED (track.label == '') OK
EXPECTED (track.language == '') OK
EXPECTED (track.cues.length == '4') OK
END OF TEST
